requirements.txt
Django==6.0.1
psycopg2-binary==2.9.10
python-decouple==3.8
Faker==21.0.0
pandas==2.2.3
requests==2.31.0
djangorestframework==3.14.0
reportlab==4.0.7
openpyxl==3.1.1
